Some logic from employees package has been moved to doctors
