+++ src/main/java/com/javateam/medicalcenter/doctors/domain/Doctor.java
@@ -1,7 +1,6 @@
 package com.javateam.medicalcenter.doctors.domain;
-
-import com.javateam.medicalcenter.employees.domain.Employee;
-import com.javateam.medicalcenter.employees.domain.Specialization;
+import com.javateam.medicalcenter.appointments.domain.Appointment;
+import com.javateam.medicalcenter.users.domain.Employee;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,18 +9,25 @@ import lombok.Setter;
 import javax.persistence.*;
 import java.math.BigDecimal;
 import java.sql.Time;
+import java.util.List;
+
 
 @Entity
-@DiscriminatorValue("Doctor")
+@Table(name="DOCTORS")
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
 public class Doctor extends Employee {
 
+    @Id
+    @GeneratedValue
+    private Long doctorId;
     private Specialization specialization;
+    @OneToMany(mappedBy = "doctor")
+    private List<Appointment> appointments;
 
-    public Doctor(Specialization specialization, String email, String name, String surname, Time dateOfBirth, Long pesel,  int hireDate, BigDecimal salary) {
+    public Doctor(Specialization specialization, String email, String name, String surname, Time dateOfBirth, Long pesel, int hireDate, BigDecimal salary) {
 
 
     }
